refactor appeals-beta entry (#5936)
